api/requirements.txt
Django==3.1.7

django-cors-headers==3.7.0
django-environ==0.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt[crypto]
djoser==2.1.0
social-auth-app-django==4.0.0

django-ckeditor==6.1.0

pytz==2022.2.1

Pillow==9.2.0
mutagen==1.45.1 

psycopg2==2.9.3 

argon2-cffi==21.3.0 

stripe==4.0.2
braintree==4.16.0

pyJWT==2.1.0

boto3==1.24.51
django-storages==1.12.2 

markdown==3.4.1
gunicorn==20.1.0
whitenoise==6.2.0

requests==2.28.1